Fix delta overflow, UPSERT sync, and OrderFlow chart tooltip
- Fix uint64 overflow in delta calculation by casting to int64
- Replace DELETE+INSERT sync with UPSERT (never deletes data)
- Remove last_sync.json dependency, track last sent time in memory
- Fix OrderFlow tooltip stuck on duplicate time keys by including date
- Replace custom Bar shape with Cell for proper mouse event handling
- Filter corrupt delta values (>1M) in frontend
- Add debug logging to dashboard data fetch

+++ backend/calculos_orderflow.py
@@ -16,7 +16,8 @@ def calcular_delta_volumen(df):
     df['volumen_bajista'] = df['tick_volume'].where(~df['es_alcista'], 0)
 
     # Delta = volumen alcista - volumen bajista
-    df['delta'] = df['volumen_alcista'] - df['volumen_bajista']
+    # Cast a int64 para evitar overflow con uint64 de MT5
+    df['delta'] = df['volumen_alcista'].astype(np.int64) - df['volumen_bajista'].astype(np.int64)
 
     return df
